# doc/swagger.yaml
definitions:
...
  crypto.KeyType:
    enum:
    - Ed25519
    - X25519
    - secp256k1
    - secp256k1-ECDSA
    - P-224
    - P-256
    - P-384
    - P-521
    - RSA
    - BLS12381G1
    - BLS12381G2
    - Dilithium2
    - Dilithium3
    - Dilithium5
    type: string
    x-enum-varnames:
    - Ed25519
    - X25519
    - SECP256k1
    - SECP256k1ECDSA
    - P224
    - P256
    - P384
    - P521
    - RSA
    - BLS12381G1
    - BLS12381G2
    - Dilithium2
    - Dilithium3
    - Dilithium5
...
  did.Document:
    properties:
      '@context': {}
      alsoKnownAs:
        type: string
      assertionMethod:
        items: {}
        type: array
      authentication:
        items: {}
        type: array
      capabilityDelegation:
        items: {}
        type: array
      capabilityInvocation:
        items: {}
        type: array
      controller:
        type: string
      id:
        description: |-
          As per https://www.w3.org/TR/did-core/#did-subject intermediate representations of DID Documents do not
          require an ID property. The provided test vectors demonstrate IRs. As such, the property is optional.
        type: string
      keyAgreement:
        items: {}
        type: array
      service:
        items:
          $ref: '#/definitions/github_com_cyware_ssi-sdk_did.Service'
        type: array
      verificationMethod:
        items:
          $ref: '#/definitions/did.VerificationMethod'
        type: array
    type: object
...
  did.VerificationMethod:
    properties:
      blockchainAccountId:
        description: for PKH DIDs - https://github.com/w3c-ccg/did-pkh/blob/90b28ad3c18d63822a8aab3c752302aa64fc9382/did-pkh-method-draft.md
        type: string
      controller:
        type: string
      id:
        type: string
      publicKeyBase58:
        type: string
      publicKeyJwk:
        allOf:
        - $ref: '#/definitions/jwx.PublicKeyJWK'
        description: must conform to https://datatracker.ietf.org/doc/html/rfc7517
      publicKeyMultibase:
        description: https://datatracker.ietf.org/doc/html/draft-multiformats-multibase-03
        type: string
      type:
        type: string
    required:
    - controller
    - id
    - type
    type: object
...
  github_com_cyware_ssi-sdk_did.Service:
    properties:
      accept:
        items:
          type: string
        type: array
      id:
        type: string
      routingKeys:
        items:
          type: string
        type: array
      serviceEndpoint:
        description: |-
          A string, map, or set composed of one or more strings and/or maps
          All string values must be valid URIs
      type:
        type: string
    required:
    - id
    - serviceEndpoint
    - type
    type: object
...
  github_com_cyware_ssi-service_pkg_service_did.CreateIONDIDOptions:
    properties:
      jwsPublicKeys:
        description: |-
          List of JSON Web Signatures serialized using compact serialization. The payload must be a JSON object that
          represents a publicKey object. Such object must follow the schema described in step 3 of
          https://identity.foundation/sidetree/spec/#add-public-keys. The payload must be signed
          with the private key associated with the `publicKeyJwk` that will be added in the DID document.
          The input will be parsed and verified, and the payload will be used to add public keys to the DID document in the
          same way in which the `add-public-keys` patch action adds keys (see https://identity.foundation/sidetree/spec/#add-public-keys).
        items:
          type: string
        type: array
      serviceEndpoints:
        description: Services to add to the DID document that will be created.
        items:
          $ref: '#/definitions/github_com_cyware_ssi-sdk_did.Service'
        type: array
    type: object
...
  pkg_server_router.BatchCreateDIDsRequest:
    properties:
      requests:
        description: Required. The list of create credential requests. Cannot be more
          than {{.Services.DIDConfig.BatchCreateMaxItems}} items.
        items:
          $ref: '#/definitions/pkg_server_router.CreateDIDByMethodRequest'
        type: array
    required:
    - requests
    type: object
  pkg_server_router.BatchCreateDIDsResponse:
    properties:
      dids:
        description: The DID documents created.
        items:
          $ref: '#/definitions/did.Document'
        type: array
    type: object
...
  pkg_server_router.CreateDIDByMethodRequest:
    properties:
      keyType:
        allOf:
        - $ref: '#/definitions/crypto.KeyType'
        description: Identifies the cryptographic algorithm family to use when generating
          this key.
      options:
        description: Options for creating the DID. Implementation dependent on the
          method.
    required:
    - keyType
    type: object
  pkg_server_router.CreateDIDByMethodResponse:
    properties:
      did:
        $ref: '#/definitions/did.Document'
    type: object
...
  pkg_server_router.GetDIDByMethodResponse:
    properties:
      did:
        $ref: '#/definitions/did.Document'
    type: object
...
  pkg_server_router.ListDIDsByMethodResponse:
    properties:
      dids:
        items:
          $ref: '#/definitions/did.Document'
        type: array
      nextPageToken:
        description: Pagination token to retrieve the next page of results. If the
          value is "", it means no further results for the request.
        type: string
    type: object
...
  pkg_server_router.ResolveDIDResponse:
    properties:
      didDocument:
        $ref: '#/definitions/did.Document'
      didDocumentMetadata:
        $ref: '#/definitions/resolution.DocumentMetadata'
      didResolutionMetadata:
        $ref: '#/definitions/resolution.Metadata'
    type: object
...
  pkg_server_router.StateChange:
    properties:
      publicKeyIdsToRemove:
        items:
          type: string
        type: array
      publicKeysToAdd:
        items:
          $ref: '#/definitions/ion.PublicKey'
        type: array
      serviceIdsToRemove:
        items:
          type: string
        type: array
      servicesToAdd:
        items:
          $ref: '#/definitions/github_com_cyware_ssi-sdk_did.Service'
        type: array
    type: object
...
  pkg_server_router.UpdateDIDByMethodRequest:
    properties:
      stateChange:
        allOf:
        - $ref: '#/definitions/pkg_server_router.StateChange'
        description: Expected to be populated when `method == "ion"`. Describes the
          changes that are requested.
    required:
    - stateChange
    type: object
  pkg_server_router.UpdateDIDByMethodResponse:
    properties:
      did:
        $ref: '#/definitions/did.Document'
    type: object
...
  resolution.Error:
    properties:
      code:
        type: string
      invalidDid:
        type: boolean
      notFound:
        type: boolean
      representationNotSupported:
        type: boolean
    type: object
  resolution.Metadata:
    properties:
      contentType:
        type: string
      error:
        $ref: '#/definitions/resolution.Error'
    type: object
...
info:
  contact:
    email: tbd-developer@squareup.com
    name: CYWARE
    url: https://github.com/cyware/ssi-service/issues
  description: 'The Self Sovereign Identity Service: Managing DIDs, Verifiable Credentials,
    and more!'
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: SSI Service API
  version: 0.0.3
paths:
...
  /v1/dids/{method}:
    get:
      consumes:
      - application/json
      description: |-
        List DIDs by method. Checks for an optional "deleted=true" query parameter, which exclusively
        returns DIDs that have been "Soft Deleted".
      parameters:
      - description: Method must be one returned by GET /v1/dids
        in: path
        name: method
        required: true
        type: string
      - description: When true, returns soft-deleted DIDs. Otherwise, returns DIDs
          that have not been soft-deleted. Default is false.
        in: query
        name: deleted
        type: boolean
      - description: Hint to the server of the maximum elements to return. More may
          be returned. When not set, the server will return all elements.
        in: query
        name: pageSize
        type: number
      - description: Used to indicate to the server to return a specific page of the
          list results. Must match a previous requests' `nextPageToken`.
        in: query
        name: pageToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_server_router.ListDIDsByMethodResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: List DIDs by method
      tags:
      - DecentralizedIdentifiers
    put:
      consumes:
      - application/json
      description: |-
        Creates a fully custodial DID document with the given method. The document created is stored internally
        and can be retrieved using the GetOperation. Method dependent registration (for example, DID web
        registration) is left up to the clients of this API. The private key(s) created by the method are stored
        internally never leave the service boundary.
      parameters:
      - description: Method
        in: path
        name: method
        required: true
        type: string
      - description: request body
        in: body
        name: request
        required: true
        schema:
          allOf:
          - $ref: '#/definitions/pkg_server_router.CreateDIDByMethodRequest'
          - properties:
              options:
                $ref: '#/definitions/github_com_cyware_ssi-service_pkg_service_did.CreateIONDIDOptions'
            type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/pkg_server_router.CreateDIDByMethodResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create a DID Document
      tags:
      - DecentralizedIdentifiers
  /v1/dids/{method}/{id}:
    delete:
      consumes:
      - application/json
      description: |-
        When this is called with the correct did method and id it will flip the softDelete flag to true for the db entry.
        A user can still get the did if they know the DID ID, and the did keys will still exist, but this did will not show up in the ListDIDsByMethod call
        This facilitates a clean SSI-Service Admin UI but not leave any hanging VCs with inaccessible hanging DIDs.
        Soft deletes a DID by its method
      parameters:
      - description: Method
        in: path
        name: method
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Soft delete a DID
      tags:
      - DecentralizedIdentifiers
    get:
      consumes:
      - application/json
      description: Gets a DID Document by its DID ID
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pkg_server_router.CreateDIDByMethodRequest'
      - description: Method
        in: path
        name: method
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_server_router.GetDIDByMethodResponse'
        "400":
          description: Bad request
          schema:
            type: string
      summary: Get a DID
      tags:
      - DecentralizedIdentifiers
    put:
      consumes:
      - application/json
      description: |-
        Updates a DID for which SSI is the custodian. The DID must have been previously created by calling
        the "Create DID Document" endpoint. Currently, only ION dids support updates.
      parameters:
      - description: Method
        in: path
        name: method
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pkg_server_router.UpdateDIDByMethodRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_server_router.UpdateDIDByMethodResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Updates a DID document.
      tags:
      - DecentralizedIdentifiers
  /v1/dids/{method}/batch:
    put:
      consumes:
      - application/json
      description: |-
        Create a batch of DIDs. The operation is atomic, meaning that all requests will
        succeed or fail. This is currently only supported for the DID method named `did:key`.
      parameters:
      - description: Method. Only `key` is supported.
        in: path
        name: method
        required: true
        type: string
      - description: The batch requests
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pkg_server_router.BatchCreateDIDsRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/pkg_server_router.BatchCreateDIDsResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Batch Create DIDs
      tags:
      - DecentralizedIdentifiers
  /v1/dids/resolver/{id}:
    get:
      consumes:
      - application/json
      description: Resolve a DID that may not be stored in this service
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_server_router.ResolveDIDResponse'
        "400":
          description: Bad request
          schema:
            type: string
      summary: Resolve a DID
      tags:
      - DecentralizedIdentifiers
...
swagger: "2.0"